Combine result of runChunkedMutation query
Initial implementation of runChunkedMutation array was returning the
result in the format of nested array. It was making hard to extract the
actual data.

This refactoring combines the result of chunk promise arrays into a
single object with key based of `createManyInputName`. Also, it aligns
to the data structure returned by the graphql query.